campus-chatbot/requirements.txt
streamlit
langchain
openai
faiss-cpu
tiktoken
beautifulsoup4
requests
python-docx
PyPDF2
pandas
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
pymysql
langchain-groq>=0.1.0
langchain>=0.1.0
langchain-community>=0.1.0
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
pymysql>=1.0.0
cryptography>=42.0.0
python-dotenv>=1.0.0
